+++ assets/javascripts/kitten/components/cards/contribution-card/index.js
@@ -1,4 +1,4 @@
-import React, { useRef } from 'react'
+import React, { useState } from 'react'
 import { CloseButton } from '../../../components/buttons/close-button'
 import { StyledContributionCard } from './styles'
 import classNames from 'classnames'
@@ -14,6 +14,7 @@ import {
   Input,
   Action,
 } from './components'
+import { Context } from './context'
 
 export const ContributionCard = ({
   className,
@@ -27,10 +28,10 @@ export const ContributionCard = ({
   borderColor,
   borderStyle,
   onClose,
+  largeInput,
   ...props
 }) => {
-  const contributionRef = useRef(null)
-
+  const [isInputEmpty, setEmptyInput] = useState(true)
   if (!show) return null
 
   return (
@@ -44,8 +45,6 @@ export const ContributionCard = ({
         '--contributionCard--border-color': borderColor,
         '--contributionCard--border-style': borderStyle,
       }}
-      ref={contributionRef}
-      role="dialog"
       {...props}
     >
       {onClose && (
@@ -62,11 +61,17 @@ export const ContributionCard = ({
         return child.props.__TYPE === 'Image' ? child : null
       })}
 
-      <div className="k-ContributionCard__gridWrapper">
-        {React.Children.map(children, child => {
-          if (!child) return null
-          return ['Image'].includes(child.props.__TYPE) ? null : child
+      <div
+        className={classNames('k-ContributionCard__gridWrapper', {
+          'k-ContributionCard__gridWrapper--largeInput': largeInput,
         })}
+      >
+        <Context.Provider value={{ isInputEmpty, setEmptyInput }}>
+          {React.Children.map(children, child => {
+            if (!child) return null
+            return ['Image'].includes(child.props.__TYPE) ? null : child
+          })}
+        </Context.Provider>
       </div>
     </StyledContributionCard>
   )
@@ -89,6 +94,7 @@ ContributionCard.defaultProps = {
   borderWidth: 2,
   imageBorderRadius: 5,
   onClose: undefined,
+  largeInput: false,
 }
 
 ContributionCard.propTypes = {
@@ -100,4 +106,5 @@ ContributionCard.propTypes = {
   borderStyle: PropTypes.string,
   borderWidth: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
   imageBorderRadius: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
+  largeInput: PropTypes.bool,
 }