+++ assets/javascripts/kitten/components/steppers/vertical-progress/index.js
@@ -41,12 +41,17 @@ const StyledContainer = styled.div`
   }
 `
 
-export const VerticalProgress = ({ children, withoutBorder, ...props }) => {
+export const VerticalProgress = ({
+  children,
+  withoutBorder,
+  className,
+  ...props
+}) => {
   return (
-    <StyledContainer className={classNames(
-      "k-VerticalProgress__wrapper",
-      {'k-VerticalProgress__wrapper--withoutBorder': withoutBorder}
-    )}
+    <StyledContainer
+      className={classNames('k-VerticalProgress__wrapper', className, {
+        'k-VerticalProgress__wrapper--withoutBorder': withoutBorder,
+      })}
     >
       <nav className="k-VerticalProgress">
         <ul role="tablist" className="k-VerticalProgress__list" {...props}>